background-replacement/requirements.txt
pillow==9.0.0
numpy<1.23.0
cffi<1.16
cython<0.30
click
